% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pasubbo}
\alias{pasubbo}
\title{Returns CDF from the AEP Distribution}
\usage{
pasubbo(x, m = 0, al = 1, ar = 1, bl = 2, br = 2)
}
\arguments{
\item{x}{(numeric) - value in the range \eqn{(-\infty, \infty)} to evaluate
the density.}

\item{m}{(numeric) - location parameter.}

\item{al, ar}{(numeric) - scale parameters. Must be in the range
\eqn{(0, \infty)}.}

\item{bl, br}{(numeric) - shape parameters. Must be in the range
\eqn{(0, \infty)}.}
}
\value{
a vector containing the values for the densities.
}
\description{
The \code{pasubbo} returns the Cumulative Distribution Function at point x
for the AEP distribution with parameters \eqn{a*}, \eqn{b*}, \eqn{m}.
}
\details{
The AEP is a exponential power distribution controlled
by five parameters, with formula:
\deqn{ f(x;a_l,a_r,b_l,b_r,m) =
\begin{cases}
\frac{1}{A} e^{- \frac{1}{b_l} |\frac{x-m}{a_l}|^{b_l} }, & x < m
\frac{1}{A} e^{- \frac{1}{b_r} |\frac{x-m}{a_r}|^{b_r} }, & x > m
\end{cases} }
with:
\deqn{A = a_lb_l^{1/b_l}\Gamma(1+1/b_l) + a_rb_r^{1/b_r}\Gamma(1+1/b_r)}
where \eqn{l} and \eqn{r} represent left and right tails, \eqn{a*} are
scale parameters, \eqn{b*} control the tails (lower values represent
fatter tails), and \eqn{m} is a location parameter.
}
