% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rasubbo}
\alias{rasubbo}
\title{Produces a random sample from a Asymmetric Power Exponential distribution}
\usage{
rasubbo(n, m = 0, bl = 2, br = 2, al = 1, ar = 1)
}
\arguments{
\item{n}{(int) - size of the sample}

\item{m}{(numeric) - location parameter.}

\item{bl, br}{(numeric) - shape parameters.}

\item{al, ar}{(numeric) - scale parameters.}
}
\value{
a numeric vector containing a random sample.
}
\description{
Generate pseudo random-number from an asymmetric power exponential distribution
using the Tadikamalla method.
}
\details{
The AEP distribution is expressed by the function:
\deqn{f(x;a_l,a_r,b_l,b_r,m) =
\begin{cases}
\frac{1}{A} e^{- \frac{1}{b_l} |\frac{x-m}{a_l}|^{b_l} }, & x < m \\
\frac{1}{A} e^{- \frac{1}{b_r} |\frac{x-m}{a_r}|^{b_r} }, & x > m
\end{cases} }
with:
\deqn{A = a_lb_l^{1/b_l}\Gamma(1+1/b_l) + a_rb_r^{1/b_r}\Gamma(1+1/b_r)}
where \eqn{m} is a location parameter, \eqn{b*} are shape parameters, \eqn{a*}
are scale parameters and \Gamma representes the gamma function.
By a suitably transformation, it is possible to use the EP distribution with
the Tadikamalla method to sample from this distribution. We basically take
the absolute values of the numbers sampled from the \code{rpower} function,
which is equivalent from sampling from a half Exponential Power distribution.
This values are then weighted by a constant expressed in the parameters.
More details are available on the package vignette and on the
function \code{rpower}.
Copyright (C) 2003-2014 Giulio Bottazzi
Copyright (C) 2020-2021 Elias Haddad
License: GPL 3+
}
\examples{
sample_gamma <- rasubbo(1000, 0, 0.5, 0.5,  1, 1)
}
