% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rgamma}
\alias{rgamma}
\title{License: GPL 3+
GSL file: randist/gamma.c}
\usage{
rgamma(n, a, b)
}
\arguments{
\item{n}{(int)}

\item{a}{(numeric)}

\item{b}{(numeric)}
}
\value{
a numeric vector containing a random sample with above parameters.
}
\description{
License: GPL 3+
GSL file: randist/gamma.c
}
\examples{
sample_gamma <- rgamma(1000, 1, 1)
}
