% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsubbotools-package.R
\docType{package}
\name{Rsubbotools-package}
\alias{Rsubbotools}
\alias{Rsubbotools-package}
\title{Rsubbotools: functions for the Subbottin and AEP distributions}
\description{
estimation of the symmetric and asymmetric power exponential, 
    a.k.a. the Subbottin family of distributions. The package is able to draw 
    samples and create parametric fits using Maximum Likelihood estimation. It 
    also provides functions for the asymmetric Laplace and skewed power 
    exponential distributions.
}
\author{
\strong{Maintainer}: Elias Haddad \email{eliasyhnetto@gmail.com} (\href{https://orcid.org/0000-0003-1863-1022}{ORCID})

Authors:
\itemize{
  \item Giulio  Bottazzi \email{bottazzi at sssup.it} (\href{https://orcid.org/0000-0003-2565-4732}{ORCID})
}

}
\keyword{internal}
