% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subbofish}
\alias{subbofish}
\title{Return the Fisher Information Matrix for the Subbotin Distribution}
\usage{
subbofish(size = 1L, b = 2, m = 0, a = 1, O_munknown = 0L)
}
\arguments{
\item{size}{numeric - number of observations (Default: 01)}

\item{b}{numeric - the exponent b (Default: 2.0)}

\item{m}{numeric - the location parameter (Default: 0.0)}

\item{a}{numeric - the scale parameter (Default: 1.0)}

\item{O_munknown}{numeric - if true assumes m known}
}
\value{
a list containing four elements:
\itemize{
\item std_error - the standard error for the parameters
\item cor_ab    - the correlation between parameters a and b
\item infmatrix - the Fisher Information Matrix
\item inv_infmatrix - the Inverse Fisher Information Matrix
}
}
\description{
Calculate the standard errors, the correlation, the
Fisher Information matrix and its inverse for a power exponential density
with given parameters
}
