% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rpower}
\alias{rpower}
\title{GSL function: gsl_ran_exppow}
\usage{
rpower(n, a, b)
}
\arguments{
\item{n}{(int) - size of the sample}

\item{a}{(numeric) - scale parameter}

\item{b}{(numeric) - shape parameter}
}
\value{
a numeric vector containing a random sample with above parameters.
}
\description{
GSL function: gsl_ran_exppow
}
\examples{
sample_gamma <- rpower(1000, 1, 1)
}
