% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rpower}
\alias{rpower}
\title{GSL function: gsl_ran_exppow}
\usage{
rpower(n, m = 0, a = 1, b = 2)
}
\arguments{
\item{n}{(int) - size of the sample.}

\item{m}{(numeric) - the location parameter.}

\item{a}{(numeric) - scale parameter.}

\item{b}{(numeric) - shape parameter.}
}
\value{
a numeric vector containing a random sample with above parameters.
}
\description{
GSL function: gsl_ran_exppow
}
\examples{
sample_gamma <- rpower(1000)
}
